Reorganized crop order of operations
