+++ RUFAS/routines/crop/nitrogen_uptake.py
@@ -127,7 +127,7 @@ def calc_log_term_of_shape_coeff(crop_type, fr_PHU_fract, fr_n_):
 # "pseudocode_SC_nitrogenuptake.docx" section 3.B.2
 #
 def calc_bio_N_opt(crop_type):
-    crop_type.bio_N_opt = crop_type.fr_N * crop_type.biomass_actual #TODO changed from prev
+    crop_type.bio_N_opt = crop_type.fr_N * crop_type.biomass_actual
 
 
 #
@@ -138,7 +138,7 @@ def calc_N_up(crop_type):
     if crop_type.bio_N_opt - crop_type.bio_N < 0:
         crop_type.N_up = 0
     else:
-        option1 = crop_type.bio_N_opt - crop_type.bio_N # TODO should be prev??
+        option1 = crop_type.bio_N_opt - crop_type.bio_N
         option2 = 4 * crop_type.fr_n3 * crop_type.dBiomass_max
 
         crop_type.N_up = min(option1, option2)
@@ -213,8 +213,8 @@ def calc_N_up_each_layer(crop_type, soil):
 def calc_N_up_z(crop_type, z):
     if crop_type.z_root == 0:
         return 0
-    term1 = crop_type.N_up / (1 - exp(-1*crop_type.beta_n))
-    term2 = 1 - exp(-1*crop_type.beta_n * z / crop_type.z_root)
+    term1 = crop_type.N_up / (1 - exp(-1 * crop_type.beta_n))
+    term2 = 1 - exp(-1 * crop_type.beta_n * z / crop_type.z_root)
     return term1 * term2
